Refactor data migration adapters to centralize pipelines, enhancing modularity and maintainability across communities, conservation status, and occurrence handlers.
