+++ boranga/components/data_migration/adapters/conservation_status/schema.py
@@ -83,41 +83,14 @@ REQUIRED_COLUMNS = [
     "processing_status",
 ]
 
-PIPELINES = {
-    # Identifiers / basics
-    "migrated_from_id": ["strip", "required"],
-    "conservation_status_number": ["strip", "blank_to_none"],
-    # Statuses / choices
-    "customer_status": ["strip", "blank_to_none", CUSTOMER_STATUS],
-    "processing_status": ["strip", "required", PROCESSING_STATUS],
-    # Simple fields
-    "conservation_criteria": ["strip", "blank_to_none"],
-    "approval_level": ["strip", "blank_to_none"],
-    "comment": ["strip", "blank_to_none"],
-    # Booleans
-    "cam_mou": ["strip", "blank_to_none", "bool"],
-    "public_consultation": ["strip", "blank_to_none", "bool"],
-    "internal_application": ["strip", "blank_to_none", "bool"],
-    "locked": ["strip", "blank_to_none", "bool"],
-    # Dates
-    "cam_mou_date_sent": ["strip", "blank_to_none", "date_iso"],
-    "public_consultation_start_date": ["strip", "blank_to_none", "date_iso"],
-    "public_consultation_end_date": ["strip", "blank_to_none", "date_iso"],
-    "review_due_date": ["strip", "blank_to_none", "date_iso"],
-    "effective_from": ["strip", "blank_to_none", "date_iso"],
-    "effective_to": ["strip", "blank_to_none", "date_iso"],
-    "listing_date": ["strip", "blank_to_none", "date_iso"],
-    "lodgement_date": ["strip", "blank_to_none", "date_iso"],
-    # User references (may be migrated to ledger ids separately)
-    "submitter": ["strip", "blank_to_none"],
-    "assigned_officer": ["strip", "blank_to_none"],
-    "assigned_approver": ["strip", "blank_to_none"],
-    "approved_by": ["strip", "blank_to_none"],
-}
+PIPELINES: dict[str, list[str]] = {}
 
 SCHEMA = Schema(
     column_map=COLUMN_MAP,
     required=REQUIRED_COLUMNS,
+    # pipelines are intentionally left empty here; source-specific bindings
+    # (eg. TPFL) are defined on the adapter module so adapters own legacy
+    # transform bindings.
     pipelines=PIPELINES,
     source_choices=None,
 )
