+++ boranga/components/data_migration/adapters/occurrence_report/schema.py
@@ -3,101 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from datetime import date
 
-import boranga.components.data_migration.mappings as dm_mappings
 from boranga.components.data_migration.adapters.schema_base import Schema
-from boranga.components.data_migration.registry import _result  # add this import
-from boranga.components.data_migration.registry import (
-    build_legacy_map_transform,
-    csv_lookup_factory,
-    dependent_from_column_factory,
-    emailuser_by_legacy_username_factory,
-    fk_lookup,
-    pluck_attribute_factory,
-    taxonomy_lookup,
-    to_int_trailing_factory,
-)
-from boranga.components.occurrence.models import OccurrenceReport
-from boranga.components.species_and_communities.models import Community, Species
-
-# build mapper that reads SHEETNO column and returns POP_ID (uses cached map)
-POP_ID_FROM_SHEETNO = dependent_from_column_factory(
-    "SHEETNO",
-    mapper=lambda dep, ctx: dm_mappings.get_pop_id_for_sheetno(
-        dep, legacy_system="TPFL"
-    ),
-    default=None,
-)
-
-TAXONOMY_TRANSFORM = taxonomy_lookup(
-    group_type_name="flora",
-    lookup_field="scientific_name",
-    source_key="TPFL",
-)
-
-SPECIES_TRANSFORM = fk_lookup(model=Species, lookup_field="taxonomy_id")
-
-COMMUNITY_TRANSFORM = fk_lookup(model=Community, lookup_field="community_name")
-
-
-def map_form_status_code_to_processing_status(value: str, ctx=None) -> str | None:
-    mapping = {
-        "NEW": OccurrenceReport.PROCESSING_STATUS_DRAFT,
-        "READY": OccurrenceReport.PROCESSING_STATUS_WITH_ASSESSOR,
-        "ACCEPTED": OccurrenceReport.PROCESSING_STATUS_APPROVED,
-        "REJECTED": OccurrenceReport.PROCESSING_STATUS_DECLINED,
-    }
-    return mapping.get(value.strip().upper()) if value else None
-
-
-def map_form_status_code_to_customer_status(value: str, ctx=None) -> str | None:
-    mapping = {
-        "NEW": OccurrenceReport.CUSTOMER_STATUS_DRAFT,
-        "READY": OccurrenceReport.CUSTOMER_STATUS_WITH_ASSESSOR,
-        "ACCEPTED": OccurrenceReport.CUSTOMER_STATUS_APPROVED,
-        "REJECTED": OccurrenceReport.CUSTOMER_STATUS_DECLINED,
-    }
-    return mapping.get(value.strip().upper()) if value else None
-
-
-# wrap the simple mapper so it fits the TransformFn signature (value, ctx) -> TransformResult
-def MAP_FORM_STATUS_CODE_TO_PROCESSING_STATUS(value, ctx):
-    return _result(map_form_status_code_to_processing_status(value, ctx))
-
-
-# keep the existing usage for dependent_from_column_factory (mapper expects dep, ctx)
-MAP_FORM_STATUS_CODE_TO_CUSTOMER_STATUS = map_form_status_code_to_customer_status
-
-# derive customer_status from the raw FORM_STATUS_CODE column
-CUSTOMER_STATUS_FROM_FORM_STATUS_CODE = dependent_from_column_factory(
-    "FORM_STATUS_CODE",
-    mapper=lambda dep_val, ctx: MAP_FORM_STATUS_CODE_TO_CUSTOMER_STATUS(dep_val, ctx),
-    default=None,
-)
-
-FK_OCCURRENCE = fk_lookup(OccurrenceReport, "migrated_from_id")
-
-OCCURRENCE_FROM_POP_ID = dependent_from_column_factory("POP_ID", FK_OCCURRENCE)
-
-# map raw RECORD_SRC_CODE -> LABEL using the provided CSV (filename required;
-# default search location is handled by load_csv_mapping)
-RECORD_SOURCE_FROM_CSV = csv_lookup_factory(
-    key_column="TERM_CODE",
-    value_column="LABEL",
-    csv_filename="DRF_LOV_RECORD_SOURCE_VWS.csv",
-    legacy_system="TPFL",
-    required=False,
-)
-
-SUBMITTER_TRANSFORM = emailuser_by_legacy_username_factory("TPFL")
-
-ROLE_TRANSFORM = build_legacy_map_transform(
-    "TPFL",
-    "ROLE (DRF_LOV_ROLE_VWS)",
-    required=False,
-)
-
-# require the "GRVL_" prefix and error if missing
-GRAVEL_TRAILING_INT = to_int_trailing_factory(prefix="GRVL_", required=True)
 
 # Header → canonical key
 COLUMN_MAP = {
@@ -128,38 +34,7 @@ REQUIRED_COLUMNS = [
     "processing_status",
 ]
 
-PIPELINES = {
-    "migrated_from_id": ["strip", "required"],
-    "Occurrence__migrated_from_id": [POP_ID_FROM_SHEETNO],
-    "species_id": ["strip", "blank_to_none", TAXONOMY_TRANSFORM, SPECIES_TRANSFORM],
-    "community_id": ["strip", "blank_to_none", COMMUNITY_TRANSFORM],
-    "lodgement_date": ["strip", "blank_to_none", "datetime_iso"],
-    "observation_date": ["strip", "blank_to_none", "date_from_datetime_iso"],
-    "record_source": ["strip", "blank_to_none", RECORD_SOURCE_FROM_CSV],
-    "customer_status": [
-        CUSTOMER_STATUS_FROM_FORM_STATUS_CODE
-    ],  # synthetic field - derived from FORM_STATUS_CODE
-    "comments": ["ocr_comments_transform"],  # synthetic field
-    "ocr_for_occ_name": [
-        OCCURRENCE_FROM_POP_ID,
-        pluck_attribute_factory("occurrence_name"),
-    ],  # synthetic field
-    "processing_status": [
-        "strip",
-        "required",
-        MAP_FORM_STATUS_CODE_TO_PROCESSING_STATUS,
-    ],
-    "submitter": ["strip", "blank_to_none", SUBMITTER_TRANSFORM],
-    # OCRObserverDetail fields
-    "OCRObserverDetail__role": ["strip", "blank_to_none", ROLE_TRANSFORM],
-    "OCRObserverDetail__observer_name": ["strip", "blank_to_none"],
-    # OCRHabitatComposition fields
-    "OCRHabitatComposition__loose_rock_percent": [
-        "strip",
-        "blank_to_none",
-        GRAVEL_TRAILING_INT,
-    ],
-}
+PIPELINES: dict[str, list[str]] = {}
 
 SCHEMA = Schema(
     column_map=COLUMN_MAP,
