+++ boranga/components/data_migration/handlers/occurrence_documents.py
@@ -96,14 +96,25 @@ class OccurrenceDocumentImporter(BaseSheetImporter):
                 r["_source"] = src
             all_rows.extend(result.rows)
 
-        # 2. Build pipelines from adapter/schema
-        pipelines = {}
-        for col, names in schema.COLUMN_PIPELINES.items():
-            from boranga.components.data_migration.registry import (
-                registry as transform_registry,
-            )
+        # 2. Build pipelines per-source by merging base schema pipelines with
+        # any adapter-provided `PIPELINES` so source-specific transforms live
+        # with the adapter while remaining runnable by the importer.
+        from boranga.components.data_migration.registry import (
+            registry as transform_registry,
+        )
 
-            pipelines[col] = transform_registry.build_pipeline(names)
+        base_column_names = schema.COLUMN_PIPELINES or {}
+        pipelines_by_source: dict[str, dict] = {}
+        for src_key, adapter in SOURCE_ADAPTERS.items():
+            src_column_names = dict(base_column_names)
+            adapter_pipes = getattr(adapter, "PIPELINES", None)
+            if adapter_pipes:
+                src_column_names.update(adapter_pipes)
+
+            built: dict[str, list] = {}
+            for col, names in src_column_names.items():
+                built[col] = transform_registry.build_pipeline(names)
+            pipelines_by_source[src_key] = built
 
         processed = 0
         errors = 0
@@ -120,7 +131,12 @@ class OccurrenceDocumentImporter(BaseSheetImporter):
             issues = []
             transformed = {}
             has_error = False
-            for col, pipeline in pipelines.items():
+            # Choose pipeline map based on row source (fallback to base)
+            src = row.get("_source")
+            pipeline_map = pipelines_by_source.get(
+                src, pipelines_by_source.get(None, {})
+            )
+            for col, pipeline in pipeline_map.items():
                 raw_val = row.get(col)
                 res = run_pipeline(pipeline, raw_val, tcx)
                 transformed[col] = res.value
