+++ boranga/components/data_migration/handlers/occurrence_reports.py
@@ -123,14 +123,33 @@ class OccurrenceReportImporter(BaseSheetImporter):
                 r["_source"] = src
             all_rows.extend(result.rows)
 
-        # 2. Build pipelines from schema
-        pipelines = {}
-        for col, names in schema.COLUMN_PIPELINES.items():
-            from boranga.components.data_migration.registry import (
-                registry as transform_registry,
-            )
+        # 2. Build pipelines per-source by merging base schema pipelines with
+        # adapter-provided `PIPELINES`. This keeps adapter-specific transforms
+        # next to the adapter implementation while the importer runs them.
+        from boranga.components.data_migration.registry import (
+            registry as transform_registry,
+        )
 
-            pipelines[col] = transform_registry.build_pipeline(names)
+        base_column_names = schema.COLUMN_PIPELINES or {}
+        pipelines_by_source: dict[str, dict] = {}
+        for src_key, adapter in SOURCE_ADAPTERS.items():
+            src_column_names = dict(base_column_names)
+            adapter_pipes = getattr(adapter, "PIPELINES", None)
+            if adapter_pipes:
+                src_column_names.update(adapter_pipes)
+
+            built: dict[str, list] = {}
+            for col, names in src_column_names.items():
+                built[col] = transform_registry.build_pipeline(names)
+            pipelines_by_source[src_key] = built
+
+        # Build a `pipelines` mapping (keys only) for merge/merge_group logic.
+        all_columns = set()
+        for built in pipelines_by_source.values():
+            all_columns.update(built.keys())
+        if not all_columns and schema.COLUMN_PIPELINES:
+            all_columns.update(schema.COLUMN_PIPELINES.keys())
+        pipelines = {col: None for col in sorted(all_columns)}
 
         processed = 0
         errors = 0
@@ -157,7 +176,12 @@ class OccurrenceReportImporter(BaseSheetImporter):
             issues = []
             transformed = {}
             has_error = False
-            for col, pipeline in pipelines.items():
+            # choose pipeline by row source
+            src = row.get("_source")
+            pipeline_map = pipelines_by_source.get(
+                src, pipelines_by_source.get(None, {})
+            )
+            for col, pipeline in pipeline_map.items():
                 raw_val = row.get(col)
                 res = run_pipeline(pipeline, raw_val, tcx)
                 transformed[col] = res.value
